some optimization fixes from Pavel
